refactor: refactor catalog api
